Preparation for the release: DailyTimeLine Screen - Clean Code rules applied
