+++ app/src/main/java/com/example/planningapp/view/partialview/_tcs/EmptyDataView.kt
@@ -1,0 +1,35 @@
+package com.example.planningapp.view.partialview._tcs
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.size
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Add
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+import com.example.planningapp.ui.theme.mainColor
+
+@Composable
+fun EmptyDataView(onAddClick: () -> Unit) {
+    Box(
+        modifier = Modifier
+            .fillMaxSize()
+            .background(Color.White),
+        contentAlignment = Alignment.Center
+    ) {
+        IconButton(onClick = onAddClick) {
+            Icon(
+                imageVector = Icons.Filled.Add,
+                contentDescription = "Add Data",
+                modifier = Modifier.size(100.dp),
+                tint = mainColor
+            )
+        }
+    }
+}
